+++ test/e2e/dataplane/tcp_gn_pod_connectivity.go
@@ -98,8 +98,8 @@ var _ = Describe("Basic TCP connectivity tests across overlapping clusters witho
 
 					err := util.Update(context.Background(), resource.ForService(framework.KubeClients[lpConfig.Cluster],
 						f.Namespace), service, func(existing *v1.Service) (*v1.Service, error) {
-						existing.Spec.Ports[0].Port = int32(lpConfig.Port)
-						existing.Spec.Ports[0].TargetPort = intstr.FromInt32(int32(lpConfig.Port))
+						existing.Spec.Ports[0].Port = lpConfig.Port
+						existing.Spec.Ports[0].TargetPort = intstr.FromInt32(lpConfig.Port)
 						return existing, nil
 					})
 					Expect(err).To(Succeed())
